+++ react-chat/package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "0.0.0",
   "type": "module",
-  "homepage": "https://romadovairina.github.io/2024-2-VK-EDU-Frontend-I-Romadova/",
+  "homepage": "https://romadovairina.github.io/2024-2-VK-EDU-Frontend-I-Romadova",
   "scripts": {
     "dev": "vite",
     "build": "vite build",
@@ -12,6 +12,7 @@
     "predeploy": "npm run build",
     "deploy": "gh-pages -d dist"
   },
+
   "dependencies": {
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
