+++ react-chat/src/api/chats/chats.js
@@ -1,0 +1,10 @@
+const getChats = () => {
+    return JSON.parse(localStorage.getItem("chats")) || [];
+  };
+  const saveChats = (chats) => {
+    localStorage.setItem("chats", JSON.stringify(chats));
+  };
+
+
+
+export{ getChats, saveChats};
